Issue Solved

+++ src/main/java/com/nichebit/resourcemanagement/service/JobCardManagementService.java
@@ -4,11 +4,13 @@ import java.util.List;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.nichebit.resourcemanagement.dto.JobCardManagementRequest;
 import com.nichebit.resourcemanagement.dto.JobCardManagementResponse;
+import com.nichebit.resourcemanagement.dto.ReturnResponse;
 import com.nichebit.resourcemanagement.entity.JobCardManagement;
 import com.nichebit.resourcemanagement.repository.JobCardManagementRepository;
 
@@ -21,28 +23,37 @@ public class JobCardManagementService {
 	@Autowired
 	private ModelMapper modelMapper;
 
-	public JobCardManagementResponse addjobcardmanagent(JobCardManagementRequest jobCardManagementRequest) {
+	ReturnResponse returnResponse = new ReturnResponse();
+
+	public ResponseEntity<?> addjobcardmanagent(JobCardManagementRequest jobCardManagementRequest) {
 		JobCardManagement jobCardManagement = this.modelMapper.map(jobCardManagementRequest, JobCardManagement.class);
 		jobCardManagementRepository.save(jobCardManagement);
-		return this.modelMapper.map(jobCardManagement, JobCardManagementResponse.class);
-
+		returnResponse.setStatus("JobCard Saved successfully.");
+		return ResponseEntity.ok(returnResponse);
 	}
 
-	public JobCardManagementResponse updatejobcardmanagent(JobCardManagementRequest jobCardManagementRequest) {
+	public ResponseEntity<?> updatejobcardmanagent(JobCardManagementRequest jobCardManagementRequest) {
 		JobCardManagement jobCardManagement = jobCardManagementRepository.findById(jobCardManagementRequest.getId())
 				.orElse(null);
 		if (jobCardManagement == null) {
-			return null;
+			returnResponse.setStatus("JobCard Not Found");
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(returnResponse);
 		}
 		jobCardManagement = this.modelMapper.map(jobCardManagementRequest, JobCardManagement.class);
 		jobCardManagementRepository.save(jobCardManagement);
-		return this.modelMapper.map(jobCardManagement, JobCardManagementResponse.class);
+		returnResponse.setStatus("JobCard Updated successfully.");
+		return ResponseEntity.ok(returnResponse);
 	}
 
-	public String deletejobcardmanagent(Long id) {
+	public ResponseEntity<?> deletejobcardmanagent(Long id) {
+		JobCardManagement jobCardManagement = jobCardManagementRepository.findById(id).orElse(null);
+		if (jobCardManagement == null) {
+			returnResponse.setStatus("JobCard Not Found");
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(returnResponse);
+		}
 		jobCardManagementRepository.deleteById(id);
-		return "delete JobCardManagement Sucessfully";
-
+		returnResponse.setStatus("Delete JobCardManagement Sucessfully");
+		return ResponseEntity.ok(returnResponse);
 	}
 
 	public List<JobCardManagementResponse> getJobCardManagents() {
@@ -60,8 +71,7 @@ public class JobCardManagementService {
 
 	public List<JobCardManagementResponse> getJobCardbyEmpID(Long EMP_ID) {
 
-		return jobCardManagementRepository.findbyEmpId(
-				EMP_ID).stream()
+		return jobCardManagementRepository.findbyEmpId(EMP_ID).stream()
 				.map(JobCardManagement -> new JobCardManagementResponse(JobCardManagement.getId(),
 						JobCardManagement.getEmpId(), JobCardManagement.getProject(), JobCardManagement.getTask(),
 						JobCardManagement.getClient(), JobCardManagement.getStartdate(), JobCardManagement.getEnddate(),
