+++ src/main/java/com/nichebit/resourcemanagement/service/TaskManagementService.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -22,8 +23,10 @@ public class TaskManagementService {
 	TaskManagementRepository taskManagementRepository;
 
 	@Autowired
-
 	ModelMapper modelmapper;
+
+
+	ReturnResponse returnResponse=new ReturnResponse();
 
 	public ResponseEntity<?> savetaskmanagement(TaskManagementRequest taskManagementRequest) {
 
@@ -46,7 +49,7 @@ public class TaskManagementService {
 		taskmanagement.setUpdationon(taskManagementRequest.getUpdationon());
 		taskmanagement.setProject(TaskProjectid);
 		taskManagementRepository.save(taskmanagement);
-		
+
 		ReturnResponse res = new ReturnResponse();
 		res.setStatus("Saved Sucessfully");
 		return ResponseEntity.ok(res);
@@ -57,7 +60,8 @@ public class TaskManagementService {
 
 		TaskManagement taskmanagement = taskManagementRepository.findById(taskManagementRequest.getId()).orElse(null);
 		if (taskmanagement == null) {
-			return ResponseEntity.notFound().build();
+			returnResponse.setStatus("Task Not Found");
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(returnResponse);
 		} else {
 			Projects TaskProjectid = new Projects();
 			TaskProjectid.setId(taskManagementRequest.getProjectid());
@@ -77,15 +81,22 @@ public class TaskManagementService {
 			taskmanagement.setUpdationon(taskManagementRequest.getUpdationon());
 			taskmanagement.setProject(TaskProjectid);
 			taskManagementRepository.save(taskmanagement);
-			ReturnResponse res = new ReturnResponse();
-			res.setStatus("Updated Sucessfully");
-			return ResponseEntity.ok(res);
+			returnResponse.setStatus("Updated Sucessfully");
+			return ResponseEntity.ok(returnResponse);
 		}
 	}
 
-	public String deletetaskmanagement(long id) {
-		taskManagementRepository.deleteById(id);
-		return "Task Deleted Successfully";
+	public ResponseEntity<?> deletetaskmanagement(long id) {
+		TaskManagement taskmanagement = taskManagementRepository.findById(id).orElse(null);
+		if (taskmanagement == null) {
+			returnResponse.setStatus("Task Not Found");
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(returnResponse);
+		} else {
+			taskManagementRepository.deleteById(id);
+			returnResponse.setStatus("Task Deleted Successfully");
+			return ResponseEntity.ok(returnResponse);
+		}
+
 	}
 
 	public List<TaskManagementResponse> gettaskmanagements(long Project_id) {
